+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 19.07.2018
+
+* [Rust] Code of Kutulu - If let in closest wanderer (CodinGame)
+
 #### 18.07.2018
 
 * [Rust] Code of Kutulu - Handle explorer function (CodinGame)